把資料庫連線配置統一放到 db.js
