+++ src/configs/db.js
@@ -1,4 +1,23 @@
 const { PrismaClient } = require('@prisma/client')
+const mysql = require('mysql2');
 const prisma = new PrismaClient()
 
+
+const db = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME,
+  port: process.env.DB_PORT,
+});
+
+db.connect(err => {
+  if (err) {
+    console.error('資料庫連線失敗:', err);
+    return;
+  }
+  console.log('資料庫連線成功');
+});
+
 module.exports = prisma
+module.exports = db
