+++ src/routes/coupon.js
@@ -1,26 +1,25 @@
 const express = require('express');
-const mysql = require('mysql2');
-const cors = require('cors');
 const router = express.Router()
+const db = require('../configs/db')
 
 
 // 資料庫連線配置//每人server不同需修改,或是創建相同account或是password
-const db = mysql.createConnection({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  port: process.env.DB_PORT,
-});
+// const db = mysql.createConnection({
+//   host: process.env.DB_HOST,
+//   user: process.env.DB_USER,
+//   password: process.env.DB_PASSWORD,
+//   database: process.env.DB_NAME,
+//   port: process.env.DB_PORT,
+// });
 
 
-db.connect(err => {
-  if (err) {
-    console.error('資料庫連線失敗:', err);
-    return;
-  }
-  console.log('資料庫連線成功');
-});
+// db.connect(err => {
+//   if (err) {
+//     console.error('資料庫連線失敗:', err);
+//     return;
+//   }
+//   console.log('資料庫連線成功');
+// });
 
 // 建立 API 端點
 router.get('/coupon', (req, res) => {
