+++ src/routes/debit.js
@@ -1,26 +1,25 @@
 const express = require('express');
-const mysql = require('mysql2');
-require('dotenv').config(); // 載入 .env 檔案
 const router = express.Router()
+const db = require('../configs/db')
 
 
 
 // 資料庫連線配置
-const db = mysql.createConnection({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  port: process.env.DB_PORT,
-});
+// const db = mysql.createConnection({
+//   host: process.env.DB_HOST,
+//   user: process.env.DB_USER,
+//   password: process.env.DB_PASSWORD,
+//   database: process.env.DB_NAME,
+//   port: process.env.DB_PORT,
+// });
 
-db.connect(err => {
-  if (err) {
-    console.error('資料庫連線失敗:', err);
-    return;
-  }
-  console.log('資料庫連線成功');
-});
+// db.connect(err => {
+//   if (err) {
+//     console.error('資料庫連線失敗:', err);
+//     return;
+//   }
+//   console.log('資料庫連線成功');
+// });
